fix: accordion closing on focus events
