+++ src/components/playground/templates/DescriptionFieldTemplate.tsx
@@ -1,0 +1,6 @@
+import React from "react"
+import {DescriptionFieldProps} from "@rjsf/utils"
+
+export const DescriptionFieldTemplate = (props: DescriptionFieldProps) => {
+  return <span className="field-description">{props.description}</span>
+}
