+++ src/components/playground/templates/index.ts
@@ -1,0 +1,3 @@
+export {ObjectFieldTemplate} from "./ObjectFieldTemplate"
+export {DescriptionFieldTemplate} from "./DescriptionFieldTemplate"
+export {FieldTemplate} from "./FieldTemplate"
